pyproject.toml
[tool.poetry]
name = "SignalCraft"
version = "0.1.0"
description = "An automated trading system using FastApi, Backtesting, and Alpaca API for data and live trading."
authors = ["Aaron Gooch <eodgooch@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"  # Use your Python version
numpy = "^1.24.0"
requests = "^2.31.0"  
duckdb = "^1.1.3"
alpaca-py = "^0.35.0"
matplotlib = "^3.10.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
ipykernel = "^6.29.5"
fastapi = {extras = ["standard"], version = "^0.115.6"}
backtesting = "^0.3.3"
uvicorn = {extras = ["standard"], version = "^0.34.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
